payment-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
stripe==7.8.0
requests==2.31.0
pydantic==2.5.0
